+++ state/animation/src/main/java/net/apartium/cocoabeans/state/animation/FixedDoubleLerpObservable.java
@@ -1,17 +1,19 @@
-package net.apartium.cocoabeans.spigot.state;
+package net.apartium.cocoabeans.state.animation;
 
+import net.apartium.cocoabeans.Mathf;
 import net.apartium.cocoabeans.state.Observable;
 import net.apartium.cocoabeans.state.Observer;
 
 import java.time.Duration;
+import java.time.Instant;
 import java.util.Collections;
 import java.util.Set;
 import java.util.WeakHashMap;
 
-public class DoubleLerpObservable implements Observable<Double>, Observer {
+public class FixedDoubleLerpObservable implements Observable<Double>, Observer {
 
     private final Observable<Double> source;
-    private final Observable<Integer> tick;
+    private final Observable<Instant> nowObservable;
     private final Set<Observer> observers = Collections.newSetFromMap(new WeakHashMap<>());
     private final double t;
     private int multiplier = 0;
@@ -23,9 +25,9 @@ public class DoubleLerpObservable implements Observable<Double>, Observer {
 
     private boolean dirty = false;
 
-    public DoubleLerpObservable(Observable<Double> source, Observable<Integer> tick, Duration duration, Duration tickRate) {
+    public FixedDoubleLerpObservable(Observable<Double> source, Observable<Instant> nowObservable, Duration duration, Duration tickRate) {
         this.source = source;
-        this.tick = tick;
+        this.nowObservable = nowObservable;
         last = this.source.get();
         start = last;
         self = last;
@@ -33,8 +35,9 @@ public class DoubleLerpObservable implements Observable<Double>, Observer {
         int totalJumps = (int) (duration.toMillis() / tickRate.toMillis());
 
         this.t = 1.0 / totalJumps;
+
         this.source.observe(this);
-        this.tick.observe(this);
+        this.nowObservable.observe(this);
     }
 
 
@@ -57,17 +60,11 @@ public class DoubleLerpObservable implements Observable<Double>, Observer {
 
         dirty = true;
 
-        self = lerp(last, current, multiplier * t);
+        self = Mathf.lerp(last, current, multiplier * t);
 
         multiplier++;
         return self;
     }
-
-    public static double lerp(double a, double b, double t) {
-        return a + (b - a) * t;
-    }
-
-
 
     @Override
     public void observe(Observer observer) {
@@ -81,7 +78,7 @@ public class DoubleLerpObservable implements Observable<Double>, Observer {
 
     @Override
     public void flagAsDirty(Observable<?> observable) {
-        if (dirty && tick == observable) {
+        if (dirty && nowObservable == observable) {
             for (Observer observer : observers)
                 observer.flagAsDirty(this);
             return;