+++ CHANGELOG.md
@@ -9,7 +9,8 @@
 - [commands] Add ContextualMapBasedParser
 - [commands] Add `VirtualCommand`
 - [minecraft] Added minecraft module
-- [spigot] Scoreboard api
+- [scoreboard] Scoreboard api for spigot minestom
+- [animation] Added Animation state
 
 ## 0.0.38
 - [commands] Add `WrappedArgumentParser` for better handling of argument parsers